indexapp/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pypdf==3.5.0
azure-identity==1.12.0
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1